+++ src/+bl12014/+ui/D141.m
@@ -13,7 +13,7 @@ classdef D141 < mic.Base
         uiStageY
         
         % {mic.ui.device.GetNumber 1x1}
-        uiMeasPointVolts
+        uiCurrent
         
     end
     
@@ -96,7 +96,7 @@ classdef D141 < mic.Base
             this.uiStageY.build(this.hFigure, dLeft, dTop);
             dTop = dTop + 15 + dSep;
             
-            this.uiMeasPointVolts.build(this.hFigure, dLeft, dTop);
+            this.uiCurrent.build(this.hFigure, dLeft, dTop);
             
         end
         
@@ -156,7 +156,7 @@ classdef D141 < mic.Base
                 'cPath',  cPathConfig ...
             );
         
-            this.uiMeasPointVolts = mic.ui.device.GetNumber(...
+            this.uiCurrent = mic.ui.device.GetNumber(...
                 'clock', this.clock, ...
                 'cName', 'measur-point-d141-diode', ...
                 'config', uiConfig, ...