+++ src/+bl12014/+ui/M143.m
@@ -16,7 +16,7 @@ classdef M143 < mic.Base
         uiStageY
                 
         % {mic.ui.device.GetNumber 1x1}
-        uiMeasPointVolts
+        uiCurrent
         
     end
     
@@ -91,7 +91,7 @@ classdef M143 < mic.Base
             this.uiStageY.build(this.hFigure, dLeft, dTop);
             dTop = dTop + 15 + dSep;
                         
-            this.uiMeasPointVolts.build(this.hFigure, dLeft, dTop);
+            this.uiCurrent.build(this.hFigure, dLeft, dTop);
             dTop = dTop + dSep;
 
             
@@ -167,7 +167,7 @@ classdef M143 < mic.Base
                 'cPath',  cPathConfig ...
             );
         
-            this.uiMeasPointVolts = mic.ui.device.GetNumber(...
+            this.uiCurrent = mic.ui.device.GetNumber(...
                 'clock', this.clock, ...
                 'cName', 'm143-current', ...
                 'config', uiConfig, ...